Move rendering to fidget-raster crate
