+++ demos/viewer/src/main.rs
@@ -13,7 +13,7 @@ use notify::{Event, EventKind, Watcher};
 use fidget::{
     gui::{Canvas2, Canvas3, CursorState, DragMode},
     gui::{View2, View3},
-    render::{GeometryPixel, ImageRenderConfig, VoxelRenderConfig},
+    raster::{GeometryPixel, ImageRenderConfig, VoxelRenderConfig},
 };
 
 use std::{error::Error, path::Path};
@@ -190,12 +190,12 @@ fn render_2d<F: fidget::eval::Function + fidget::render::RenderHints>(
 
         Mode2D::Sdf => {
             let tmp = config.run(shape).unwrap();
-            fidget::render::effects::to_rgba_distance(tmp, config.threads)
+            fidget::raster::effects::to_rgba_distance(tmp, config.threads)
         }
 
         Mode2D::Debug => {
             let tmp = config.run(shape).unwrap();
-            fidget::render::effects::to_debug_bitmap(tmp, config.threads)
+            fidget::raster::effects::to_debug_bitmap(tmp, config.threads)
         }
     };
     let (data, _) = out.take();
