+++ fidget-core/src/gui/mod.rs
@@ -21,7 +21,7 @@ use serde::{Deserialize, Serialize};
 ///
 /// ```
 /// # use nalgebra::{Vector2, Point2};
-/// # use fidget_core::render::{View2};
+/// # use fidget_core::gui::{View2};
 /// let view = View2::from_center_and_scale(Vector2::new(5.0, 5.0), 1.0);
 ///
 /// //   -------d-------
