+++ fidget-raster/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "fidget-raster"
+description = "Bitmap and heightmap rendering for complex closed-form implicit surfaces"
+readme = "../README.md"
+
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+authors.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+fidget-core = { path = "../fidget-core", version = "=0.4.0" }
+workspace-hack = { version = "0.1", path = "../workspace-hack" }
+
+nalgebra.workspace = true
+ordered-float.workspace = true
+rand.workspace = true
+rayon.workspace = true
+zerocopy.workspace = true
